requirements.txt
streamlit
opencv-python
numpy
tensorflow
scikit-image
matplotlib
Pillow
